requirements.txt
flask
flask-sqlalchemy
flask-bootstrap
flask-login
flask-restful
wtforms-sqlalchemy
wtforms
flask-migrate
flask-script
flask-wtf
psycopg2
pyjwt 
ipaddress
flask-wtf-ext
wtforms_components
email_validator
openpyxl
setuptools
Werkzeug
SQLAlchemy
alembic
python-dotenv